school_api/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
CacheControl==0.12.10
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
distlib==0.3.4
distro==1.7.0
Django==4.0.3
django-automatic-crud==1.2.0
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.1.1
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.3
mysqlclient==2.1.0
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.3
pep517==0.12.0
Pillow==9.0.1
progress==1.6
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
PySocks==1.7.1
pytoml==0.1.21
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
retrying==1.3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
zope.event==4.5.0
zope.interface==5.4.0
